Adding readOnly parameter to graphdriver Create method
Since the layer store was introduced, the level above the graphdriver
now differentiates between read/write and read-only layers.  This
distinction is useful for graphdrivers that need to take special steps
when creating a layer based on whether it is read-only or not.
Adding this parameter allows the graphdrivers to differentiate, which
in the case of the Windows graphdriver, removes our dependence on parsing
the id of the parent for "-init" in order to infer this information.

This will also set the stage for unblocking some of the layer store
unit tests in the next preview build of Windows.

cherry-pick from: https://github.com/moby/moby/pull/20525
we need this pr to solve the conflicts when cherry-picking
overlay2.

Signed-off-by: Stefan J. Wernli <swernli@microsoft.com>
Signed-off-by: Lei Jitang <leijitang@huawei.com>
(cherry picked from commit ef5bfad3210a9e9c8b761f2c11c0c6289490ebff)