The following syscalls should not be blocked by seccomp
Cherry-picked From:
https://github.com/projectatomic/docker/commit/0e125524a066fd797a0398f337beab716b11892a

Changes:

* use `--system-container` to unblocked the following syscalls when we start system container

Capabilities block these syscalls.

mount, umount2, unshare, reboot and name_to_handle_at are all needed to
run systemd as pid1 in a container, they work fine with sys_admin disabled
and have functionality in the kernel that is available to a non privileged
process.  There is no easy way to discover which syscalls are blocked, so
we end up more likely with the user doing a --privileged.

With UserNamespace we want to allow users to potentially setup unshare additional
namespaces.

man reboot
...
   Behavior inside PID namespaces
       Since Linux 3.4, when reboot() is called from a PID namespace (see
       pid_namespaces(7)) other than the initial PID namespace, the effect
       of the call is to send a signal to the namespace "init" process.
       LINUX_REBOOT_CMD_RESTART and LINUX_REBOOT_CMD_RESTART2 cause a SIGHUP
       signal to be sent.  LINUX_REBOOT_CMD_POWER_OFF and
       LINUX_REBOOT_CMD_HALT cause a SIGINT signal to be sent.

Signed-off-by: Dan Walsh <dwalsh@redhat.com>
Signed-off-by: Antonio Murdaca <runcom@redhat.com>
Signed-off-by: Wang Long <long.wanglong@huawei.com>