Relative symlinks don't work with --device argument
This fix tries to address the issue raised in #22271 where
relative symlinks don't work with --device argument.

Previously, the symlinks in --device was implemneted (#20684)
with `os.Readlink()` which does not resolve if the linked
target is a relative path. In this fix, `filepath.EvalSymlinks()`
has been used which will reolve correctly with relative
paths.

An additional test case has been added to the existing
`TestRunDeviceSymlink` to cover changes in this fix.

backport from: https://github.com/docker/docker/pull/22272

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>
Signed-off-by: Yuanhong Peng <pengyuanhong@huawei.com>