Fix escape-keys by preserving input if invalid
Currently, using a custom detach key with an invalid sequence, eats a
part of the sequence, making it weird and difficult to enter some key
sequence.

This fixes by keeping the input read when trying to see if it's the key
sequence or not, and "writing" then is the key sequence is not the right
one, preserving the initial input.

(backport from https://github.com/moby/moby/pull/22943)

Signed-off-by: Vincent Demeester <vincent@sbr.pm>
Signed-off-by: Wentao Zhang <zhangwentao234@huawei.com>
(cherry-pick from 0fb6190243d6101f96283e487cd4911142a05483)