Fix docker rename with linked containers
This fix tries to address issue raised in #23973 where the
`docker rename` does not update namedIndex and linkIndex,
thus resulting in failures when the linked containers are
referenced later on.

This fix updates the namedIndex and linkIndex during the
`docker rename` and fixes the issue.

An integration test has been added to cover the changes
in this fix.

This fix fixes #23973.

cherry-pick from https://github.com/moby/moby/pull/24020
fix DTS2017070607365

Signed-off-by: Chen Min <chenmin46@huawei.com>
Signed-off-by: Yong Tang <yong.tang.github@outlook.com>