Add support for --attach-output
This is a debug feature, with --attach-output <container-id>,
you can attach output from another container, so all output
of <container-id> will be printed to the stdout of current
container.

This is used with debug tools like GDB, some debuging need
to send certain commands to processes which are in other
containers, and these commands will print information to
these processes' stdout, with this feature, you can see
these outputs in your current debug container's stdout.

Signed-off-by: Qiang Huang <h.huangqiang@huawei.com>