Return an empty stats if the container is restarting
In case, a container is restarting indefinitely running
"docker stats --no-stream <restarting_container>" is suspended.

To fix this, the daemon makes sure the container is either not
running or restarting if `--no-stream` is set to true and if so
returns an empty stats.

Should fix #27772.

fix DTS2017062812251
backport from https://github.com/moby/moby/pull/28026 without conflicts.

Signed-off-by: Boaz Shuster <ripcurld.github@gmail.com>
Signed-off-by: Yuanhong Peng <pengyuanhong@huawei.com>