Do not remove containers from memory on error
Before this, if `forceRemove` is set the container data will be removed
no matter what, including if there are issues with removing container
on-disk state (rw layer, container root).

In practice this causes a lot of issues with leaked data sitting on
disk that users are not able to clean up themselves.
This is particularly a problem while the `EBUSY` errors on remove are so
prevalent. So for now let's not keep this behavior.

Cherry-pick from https://github.com/moby/moby/pull/31012

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
Signed-off-by: Wentao Zhang <zhangwentao234@huawei.com>
(cherry picked from commit 54dcbab2)

Conflicts:
	daemon/delete.go
	daemon/graphdriver/aufs/aufs.go
	daemon/graphdriver/btrfs/btrfs.go
	daemon/graphdriver/devmapper/driver.go
	daemon/graphdriver/overlay/overlay.go
	pkg/mount/mount.go