Fix remove none image when HoldOff the image
Description:
We have already added `ReferenceCount` for image in MR
http://code.huawei.com/docker/docker/merge_requests/405

But it has following issue:
1. The origin RefCount is initialized by 1, and decreased while deleting
it. This means if more than one DELETE requests come, the image could be
deleted.

2. When HoldOn an image, the origin behavior is to increase all the parents
of the image. This is ugly.

So this commit solve 2 issues:
1. Delete operation will not decrease RefCount, if RefCount>0, Delete
operation will fail.
2. HoldOn an image, only increase RefCount of itself and its parent.
Just like the layer referenceCount behavior.

Signed-off-by: Wentao Zhang <zhangwentao234@huawei.com>