Fix save & rmi image competition
Description:
The orgin solution is when saving an image, will set a status to a
image. If rmi operation comes, it will report error.

But this is ugly.

New solution:
We have Refcount for each image, when we begin to save images, Holdon
all the images first. Then delete won't success.

If failed to holdon, this means the deletion is ahead of saving, so save
operation will report an holdon error.

Fix DTS:
DTS2017051508777

Signed-off-by: Wentao Zhang <zhangwentao234@huawei.com>