Ensure log driver is not nil.
When getting log driver, it'll first check if container's log driver
is nil. If not nil, it will return the log driver. Between judgement
and the return, there is enough time for log driver to be nil, because
function container.IsRunning() will wait lock if container is exiting
or restarting.

Fix DTS2017051900715
Close issue #289

Signed-off-by: Fengtu Wang <wangfengtu@huawei.com>