Print infomation of base image when loading partial image.
When loading partial image, we hope to load all it's depend
images. Option -p will find it's dependent image loop until
found dependent image not exists in local or no more dependent
image. It will print the dependent image found if it's built
with option --no-parent, then shell script can pull it's
depend image and load it. If it's not built with option
--no-parent, it'll print nothing. Printed information like:

```
  $ docker load -i noparent.tar -p
  ToBePulled: busybox:latest@sha256:00f017a8c2a6e1fe2ff...
```

Signed-off-by: Fengtu Wang <wangfengtu@huawei.com>