devmapper: remove broken device when start daemon
When use `kill -9 docker-daemon-pid` to stop docker, It may casue that
the status of devicemapper devices inconsisitent. eg, we can see the
device info use `dmsetup status` command, but the device does not exist
in `/dev/mapper` directory.

At next time docker start, it will start failed or load container
failed.

This patch add a check when docker start in `initDevmapper` function.
First, we find all devmapper devices in the system. Then If the device
name has a prefix with `docker-{major}:{minor}-{inode}` and the length
is equal to zero, we use `devicemapper.RemoveDevice` remove the device.

Fix: DTS2017051611286

Signed-off-by: Wang Long <long.wanglong@huawei.com>