Ignore ToDisk error in StateChanged
If container is started normally, but failed to save
state to disk due to error "no space left on device",
then container can not be stopped, because container's
infomation in libcontainerd is cleaned up when error
occurred(so it can not process event 'exit').

We can ignore ToDisk error in StateChanged, it only
change status of container. Status is correct if docker
daemon not restart, because right status exists in memory.
If docker daemon restart, it will restore these status
using status of containerd, so status is also correct.

This fix can break consistency of status in memory and
disk, but considering there is no space in disk, it is
not a big problem in this situation. Status in disk can
recover if disk have space and if ToDisk is written
again(for example, status changed).

Fix issuse #322
Fix DTS2017062011340

Signed-off-by: Fengtu Wang <wangfengtu@huawei.com>