Return exit status immediately if container is restarting
If tring to wait and get exit code using command
"docker wait" when container is waiting for restart,
we should return exit code of last exit, but not
waiting for container to start and then exit.
Because event "start" can also break waiting, so
when getting exit code, you will got zero as
container's status is changed to running just now.

Fix DTS2017062910718

Signed-off-by: Fengtu Wang <wangfengtu@huawei.com>