accel: Allow create non-persistent accelerator by REST-API
This patch is used for K8S integration. The K8S assume resources
are available when container is stopped, but the persistent
accelerators will not be released until container is deleted.
So we remove IsPersistent checking in daemon to allow user like
K8S to create non-persistent accelerator by REST-API.

This modify do not affect accelerators created by cli options,
which are still persistent.

Signed-off-by: majiuyue 00385406 <majiuyue@huawei.com>