Bugfix: Host netns file "default" not mounted
This commit is fixing a bug: if the "/var/run/docker/netns/default" is
unmounted, and there is already a container running with host network
mode, then restarting docker daemon can't mount the "default" netns file
again which lead to failure of creating new container with host network
mode.

To reproduce this bug:

1. starting docker daemon with "--live-restore" flag
2. run a container with "--net host"
3. umount default file with "$ umount /var/run/docker/netns/default"
4. restart docker daemon with same flags.
5. run a new container with "--net host" WILL fail, and you will find
the "default" netns file isn't mounted.

Signed-off-by: Zhang Wei <zhangwei555@huawei.com>